requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
djangorestframework==3.15.2
tzdata==2025.2